+++ CHANGELOG.md
@@ -38,6 +38,7 @@ is roughly based on [Keep a Changelog], and this project tries to adheres to
 - DPKG metadata is no longer incorrectly deleted for container images (bug
   exposed in v0.9.1)
 - `merge_sam` no longer panics on SRA qnames that are parseable but lack an explicit read side
+- `phase` no longer truncates variants table file on malformed or corrupted variants matrix files
 
 ## [0.10.0] - 2026-06-05